Sketch of FS service API
